Split log and UI output
Log output is considered anything with a level -- so, info, warning,
and error messages are log output. Info level currently goes to stdout,
making it a requirement to pass -q to spin in order to get something
that can be piped to jq or equivalent.

I'd like to replace this entire cli.Ui bit of code because it seems
like it's just causing more problems than it's solving, but for now,
split the outputs along stdout and stderr FDs and move info, error, and
warning all to the stderr output.